DF-8024 warning fix
Former-commit-id: 27547eca44ba0d3d1e3010f8d741acc0d9b73800

+++ Sources/Internal/Render/RHI/DX9/rhi_QueryBufferDX9.cpp
@@ -112,7 +112,7 @@ dx9_Check_Query_Results(QueryBufferDX9_t* buf)
             uint32 resultIndex = buf->pendingQueries[q].second;
             if (cmd[q].retval == S_OK)
             {
-                if (resultIndex < buf->results.size())
+                if (resultIndex < uint32(buf->results.size()))
                     buf->results[resultIndex] = results[q];
 
                 QueryDX9Pool.push_back(buf->pendingQueries[q].first);
@@ -173,7 +173,7 @@ dx9_QueryBuffer_Value(Handle handle, uint32 objectIndex)
 
     dx9_Check_Query_Results(buf);
 
-    if (objectIndex < buf->results.size())
+    if (objectIndex < uint32(buf->results.size()))
     {
         return buf->results[objectIndex];
     }
